spring, decay, timing
