+++ src/reanimated2/animation/commonTypes.ts
@@ -1,0 +1,9 @@
+export interface Animation {
+
+}
+
+export interface AnimationConfig {
+  
+}
+
+export type AnimationCallback = () => unknown;
