+++ src/reanimated2/animation/index.ts
@@ -1,1 +1,6 @@
 export * from './animations';
+export * from './commonTypes';
+// export * from './utils';
+export * from './timing';
+export * from './spring';
+export * from './decay';
